+++ flinkboot-test/src/main/java/module-info.java
@@ -4,7 +4,13 @@ module io.github.sekelenao.flinkboot.test {
     requires org.junit.jupiter.params;
 
     exports io.github.sekelenao.flinkboot.test.api;
+    exports io.github.sekelenao.flinkboot.test.api.assertion;
+    exports io.github.sekelenao.flinkboot.test.api.assertion.type;
 
     opens io.github.sekelenao.flinkboot.test.api;
+    opens io.github.sekelenao.flinkboot.test.api.assertion;
+    opens io.github.sekelenao.flinkboot.test.api.assertion.type;
     opens io.github.sekelenao.flinkboot.test.internal to org.junit.platform.commons;
 }
+
+